OOP version 2 commit
