+++ .idea/workspace.xml
@@ -1,18 +1,40 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
   <component name="ChangeListManager">
-    <list default="true" id="005e55b5-02b7-4fdd-beef-48a417d61102" name="Changes" comment="" />
+    <list default="true" id="005e55b5-02b7-4fdd-beef-48a417d61102" name="Changes" comment="OOP version 1 commit">
+      <change afterPath="$PROJECT_DIR$/src/javacore/oop/InputPublication.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/javacore/oop/ValidationUnits.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/javacore/oop/Library.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/javacore/oop/Library.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/javacore/oop/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/javacore/oop/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/javacore/oop/Publication.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/javacore/oop/Publication.java" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Class" />
+      </list>
+    </option>
+  </component>
   <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="javacoreBasic" />
+      </map>
+    </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
   </component>
-  <component name="ProjectColorInfo"><![CDATA[{
-  "associatedIndex": 2
-}]]></component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 2
+}</component>
   <component name="ProjectId" id="31TI4Ril2Brppla8FAzWF9EyTCg" />
   <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
   <component name="ProjectViewState">
@@ -22,7 +44,7 @@
   "keyToString": {
     "RunOnceActivity.OpenProjectViewOnStart": "true",
     "RunOnceActivity.ShowReadmeOnStart": "true",
-    "git-widget-placeholder": "master"
+    "git-widget-placeholder": "javaCoreOOP"
   }
 }]]></component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
@@ -34,6 +56,19 @@
       <option name="presentableId" value="Default" />
       <updated>1755541446070</updated>
     </task>
+    <task id="LOCAL-00001" summary="OOP version 1 commit">
+      <option name="closed" value="true" />
+      <created>1756033091762</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1756033091762</updated>
+    </task>
+    <option name="localTasksCounter" value="2" />
     <servers />
+  </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="OOP version 1 commit" />
+    <option name="LAST_COMMIT_MESSAGE" value="OOP version 1 commit" />
   </component>
 </project>
